Update README.md
Updated LivingDoc creation command.
Due to .NET tooling limitations, it is not possible trigger LivingDoc generation automatically after a test run using only the .csproj file and dotnet test. The recommended and most reliable way is to use a script that runs both the test and LivingDoc generation commands in sequence.
run-tests-with-livingdoc.bat is therefore in root.